Had the initial idea of using DFS and just going to the right side only - didn't work.
Optimal solution is to use BFS and then check for the most right side unit and keep a hold of the most right side node by continually updating the right side node and using the fact that  we're using our deque left to right to always use the most right_sided value as our update. This is reinforced by the if node check.